+++ spec/spec_helper.rb
@@ -60,13 +60,14 @@ RSpec.configure do |config|
 
     next unless example.metadata[:type] == :pro
 
-    Karafka::Pro::Loader.pre_setup(Karafka::App.config)
+    Karafka::Pro::Loader.pre_setup_all(Karafka::App.config)
   end
 
   config.after do
     Karafka::App.routes.clear
     Karafka.monitor.notifications_bus.clear
     Karafka::App.config.internal.routing.activity_manager.clear
+    Karafka::Processing::InlineInsights::Tracker.clear
   end
 end